More OOP
Abstracted out the functionalities of the chatbot, the task list and the storage function.
